Generic BeaconRssiMeasurement

+++ blazor-application/Data/DataPoint.cs
@@ -4,9 +4,9 @@ public sealed class DataPoint
 {
     public ClassLabel Label { get; }
     
-    public int[] Rssis { get; }
+    public double[] Rssis { get; }
 
-    public DataPoint(ClassLabel label, int[] rssis)
+    public DataPoint(ClassLabel label, double[] rssis)
     {
         Label = label;
         Rssis = rssis;
@@ -14,11 +14,11 @@ public sealed class DataPoint
 
     public static double Distance(DataPoint p1, DataPoint p2) => Distance(p1.Rssis, p2.Rssis);
 
-    public static double Distance(int[] rssis, DataPoint p) => Distance(rssis, p.Rssis);
+    public static double Distance(double[] rssis, DataPoint p) => Distance(rssis, p.Rssis);
     
-    public static double Distance(DataPoint p, int[] rssis) => Distance(p.Rssis, rssis);
+    public static double Distance(DataPoint p, double[] rssis) => Distance(p.Rssis, rssis);
 
-    public static double Distance(int[] rssis1, int[] rssis2)
+    public static double Distance(double[] rssis1, double[] rssis2)
     {
         if (rssis1.Length != rssis2.Length)
         {
